requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
django-modeltranslation==0.17.5
djangorestframework==3.12.4
drf-yasg==1.20.0
flower==1.0.0
gunicorn==20.1.0
humanize==3.12.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.0
MarkupSafe==2.0.1
oauthlib==3.1.1
packaging==21.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.2.0
pyparsing==2.4.7
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.1
uritemplate==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
